add group medicine list

+++ router/medicine-update.js
@@ -89,6 +89,7 @@ router.post('/medicine-update/:medicine_id', async (ctx) => {
     let startsDate = ctx.request.body['starts_date'];
     let period = ctx.request.body['period'];
     let medicineType = ctx.request.body['medicine_type'];
+    let groupId = ctx.request.body['group_id'];
 
     // 任意項目
     let medicineImage = "";
@@ -131,12 +132,13 @@ router.post('/medicine-update/:medicine_id', async (ctx) => {
     ]);
     let validationTakeTime = await app.validationTakeTime(takeTime);
     let validationMedicineType = await app.validationMedicineType(medicineType);
+    let validationGroupId = await app.validationGroupId(groupId, userId);
 
-    if (validationMedicine.result && validationTakeTime && validationMedicineType && uploadImageFlag) {
+    if (validationMedicine.result && validationTakeTime && validationMedicineType && validationGroupId && uploadImageFlag) {
         // Update Medicine
         let sql = 'UPDATE medicine SET medicine_name = ?, hospital_name = ?, number = ?, starts_date = ?, ' +
-            'period = ?, type_id = ?, image = ?, description = ? WHERE medicine_id = ?';
-        await connection.query(sql, [medicineName, hospitalName, number, startsDate, period, medicineType, medicineImage, description, medicineId]);
+            'period = ?, type_id = ?, group_id = ?, image = ?, description = ? WHERE medicine_id = ?';
+        await connection.query(sql, [medicineName, hospitalName, number, startsDate, period, medicineType, groupId, medicineImage, description, medicineId]);
 
         // Delete TakeTime
         sql = 'DELETE FROM medicine_take_time WHERE medicine_id = ?';
@@ -162,11 +164,13 @@ router.post('/medicine-update/:medicine_id', async (ctx) => {
         if (startsDate !== '') session.old.starts_date = startsDate;
         if (period !== '') session.old.period = period;
         if (medicineType !== '') session.old.type_id = medicineType;
+        if (groupId !== '') session.old.group_id = groupId;
         if (description !== '') session.old.description = description;
         if (!uploadImageFlag) session.error.medicine_image = '1MB以内のJPEG・JPG・PNG・ファイルを選択してください';
 
         if (!validationTakeTime) session.error.take_time = '飲む時間が正しく選択されていません';
         if (!validationMedicineType) session.error.medicine_type = '種類が正しく選択されていません';
+        if (!validationGroupId) session.error.group_id = 'グループが正しく選択されていません';
 
         session.error.message = '薬情報の更新に失敗しました';
